Changes in Personal Cabinet, Changes in Admin Panel

+++ PetCare.Domain/Models/Pet/PetDetails.cs
@@ -6,8 +6,8 @@ public class PetDetails
 {
     [Key]
     public Guid Id { get; set; }
-    public float Weight { get; set; }
-    public float Height { get; set; }
+    public double Weight { get; set; }
+    public double Height { get; set; }
     public DateTime? CreatedAt { get; set; }
     
     // References
